- FileSystem implementation + test   - FileSystem will allocate the memory for any File opened.   - Calling CloseFile will release that memory - Also added more tests to the smallVector to make sure the internal pointers can be void* without problems.

+++ Source/Core/Core.h
@@ -25,7 +25,7 @@
 
 typedef unsigned char uint8;
 typedef unsigned short uint16;
-typedef unsigned int uint32;
+typedef unsigned long uint32;
 typedef unsigned long long uint64;
 
 typedef int int32;
